Revert "Authentication ray -- finished authentication "
