+++ backend/middleware/model-middleware.js
@@ -1,7 +1,7 @@
 const User = require('../models/users');
 
 /**
- * This middleware function is meant to be registered BEFORE the route handlers (see server.js)
+ * This middleware function is meant to be registered BEFORE the route handlers (see index.js)
  * This sets up a connection to the database. We modify the request object by tacking on the
  * models and disconnect function. Any FUTURE middleware / route handler thus has access to
  * those models / disconnect function by virtue of the fact that the request object is the same
