+++ backend/package.json
@@ -11,7 +11,6 @@
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
     "knex": "^1.0.4",
     "mysql": "^2.18.1"
   },
