+++ backend/routes/users.js
@@ -1,7 +1,6 @@
 const { application } = require('express');
 const express = require('express');
-const userController = require('../controllers/users.js');
-const userModel = require('../models/users.js');
+const users = require('../models/users.js');
 
 
 /**
@@ -12,82 +11,43 @@ const userModel = require('../models/users.js');
  */
 const router = express.Router();
 
-//ROUTES FOR USERS
 
-//GET /users -- returns all users for a given search perameter
-//Can search based off of firstName, lastName, email, or ID
-router.get('/', async(req, res, next) => {
-    try{
-        const result = await userController.getUsers(req.query.firstName, req.query.lastName, req.query.email, req.query.id);
-        res.status(200).json(result);
-    }catch(err){
-        console.error('Failed to get user info:', err);
-        res.status(500).json({ message: err.toString() });
-    }
-    next();
-});
-
-//GET /users/self -- returns the current user information (basically just a call to the GET /users route with the current user's ID)
-router.get('/self', async(req, res, next) => {
-    try{
-        const userDoesExist = await userController.userDoesExist(null,null,null,req.user.id);
-        if(userDoesExist){
-            const result = await userController.getUsers(null, null, null,req.user.id);
-            res.status(200).json(result[0]);
-        }else{
-            res.status(500).json({ message: 'Failed to get user info' });
-        }
-    }catch(err){
-        console.error('Failed to get user info:', err);
-        res.status(500).json({ message: err.toString() });
-    }
-    next();
-});
-
-//PUT /users/self -- updates the user's information
-router.put('/self', async (req, res, next) => {
+router.post('/register', async (req, res, next) => {
     try {
-        const result = await userController.updateUser(req.user.id, req.body.firstName, req.body.lastName, req.body.email, req.body.password, req.body.bio);
-        if (result.error === "User account Does not exist") {
-            res.status(400).json({message: result.error});
-        } else if (result.error === "No changes entered") {
-            res.status(400).json({message: result.error});
-        } else if (result.error === "Changes conflict with existing user") {
-            res.status(400).json({message: result.error});
+        const body = req.body;
+        console.log(body);
+    //    const result = await req.models.user.createNewUser(body.email, body.password);
+       // const result = await req.models.createNewUser(req.body.firstName, req.body.lastName, req.body.email, req.body.password
+      //calls the createNewUser function in the users.js file of the models folder and return the result
+        const result = await users.createNewUser(body.firstName, body.lastName, body.email, body.password);
+        console.log("Result of createNewUser: ", result);
+        if(result.error === "Invalid email"){
+            console.log("Invalid email");
+            res.status(400).json({message: "Invalid email"});
+        } else if(result.error === "User already exists"){
+            console.log("User already exists");
+            res.status(400).json({message: "User already exists"});
         } else {
-            const output = await userController.getUsers(null, null, null, req.user.id);
-            res.status(200).json(output[0]);
+            console.log("User created");
+            res.status(201).json(result);
         }
+
     } catch (err) {
-        console.error('Failed to update user info:', err);
+        console.error('Failed to create new user:', err);
         res.status(500).json({ message: err.toString() });
     }
-});
 
-//DEL /users/self -- deletes the user's information
-router.delete('/self', async (req, res, next) => {
-    try{
-        const userDoesExist = await userController.userDoesExist(null, null, null, req.user.id);
-        if(userDoesExist) {
-            const result = await userController.deleteUser(req.user.id);
-            res.status(204).json({ message: 'User deleted' });
-        } else {
-            res.status(400).json({ message: 'User does not exist' });
-        }
-    } catch (err) {
-        console.error('Failed to delete user info:', err);
-        res.status(500).json({ message: err.toString() });
-    }
-});
+    next();
+})
 
 
-//This route is superflous once the GET /users route is implemented
+//get id route. Given the email of a of a user, return the id of that user
 router.get('/id/:email', async (req, res, next) => {
     try {
         const email = req.params.email;
 
 
-        const result = await userModel.getIDFromEmail(email);
+        const result = await users.getIDFromEmail(email);
         if(result === -1){
             res.status(300).json({message: "User not found"});
         } else {
@@ -101,6 +61,25 @@ router.get('/id/:email', async (req, res, next) => {
     next();
 })
 
-
+//get user by id
+router.get('/info/:id', async(req, res, next) => {
+    try{
+        const id = req.params.id;
+        console.log("Id is: " + id);
+        const result = await users.getUserById(id);
+        //console.log("Result is: " + result.json);
+        if(result.length > 0) {
+            res.status(200).json(result);
+        } else {
+            res.status(400).json({
+                message: "User not found"
+            });
+        }
+    }catch(err){
+        console.error('Failed to get user info:', err);
+        res.status(500).json({ message: err.toString() });
+    }
+    next();
+});
 
 module.exports = router;
