+++ frontend/src/api/userRepository.js
@@ -14,7 +14,7 @@ export class UserRepository {
     async register(firstname, lastname, email, pass) {
         const errors = { success: false };
         console.log('about tot try to post1');
-        const { status } = await axios.post(URL + "/register", {firstName: firstname, lastName: lastname, email: email, password: pass})
+        const { status } = await axios.post(URL + "/users/register", {firstName: firstname, lastName: lastname, email: email, password: pass})
             .then(function (response) {
                 console.log('this is the response: ');
                 errors.success = true;
@@ -41,7 +41,7 @@ export class UserRepository {
     async login(email, pass) {
         const errors = { success: false };
         console.log('about tot try to post2');
-        const { status } = await axios.post(URL + "/login", {email: email, password: pass});
+        const { status } = await axios.post(URL + "/session/login", {email: email, password: pass});
         console.log("The status is: " + status);
         if (status <= 201) {
         errors.success = true;
