+++ package.json
@@ -1,5 +1,8 @@
 {
   "dependencies": {
     "bcrypt": "^5.0.1"
+  },
+  "devDependencies": {
+    "@iconify/react": "^3.2.1"
   }
 }
